refactor: Restructure las carpetas del proyecto
