+++ app/components/dashbord/AccomodationCardComponent.jsx
