+++ app/components/dashbord/TourismCardComponent.jsx
