+++ app/components/generals/SearchInputComponent.jsx
